refactor: organize download-related types and error handling utilities
